pyproject.toml
[tool.poetry]
name = "airflow-trigger-multiple-dagrun"
version = "1.0.0"
description = "Plugin to trigger and monitor multiple dag runs"
authors = ["Jaya Bharath <bharath21196@gmail.com>"]
readme = "README.md"
packages = [{include = "airflow_trigger_multiple_dagrun"}]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7.13"
apache-airflow = "^2.6.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
